# public/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Client Management API
  description: API REST para gerenciamento de clientes com validação de CPF
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: Authentication
    description: Autenticação e autorização
  - name: Clients
    description: Gerenciamento de clientes

paths:
  /api/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Autentica um cliente e retorna um token de acesso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: joao.silva@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: 1|abc123def456...
                      token_type:
                        type: string
                        example: Bearer
                      client:
                        $ref: '#/components/schemas/Client'
        '422':
          description: Dados de validação inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /api/logout:
    post:
      tags:
        - Authentication
      summary: Logout
      description: Invalida o token de acesso atual
      security:
        - sanctum: []
      responses:
        '200':
          description: Logout realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthenticated'

  /api/me:
    get:
      tags:
        - Authentication
      summary: Get authenticated user
      description: Retorna os dados do cliente autenticado
      security:
        - sanctum: []
      responses:
        '200':
          description: Dados do cliente autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthenticated'

  /api/clients:
    get:
      tags:
        - Clients
      summary: List clients
      description: Lista todos os clientes com filtros opcionais e paginação
      security:
        - sanctum: []
      parameters:
        - name: name
          in: query
          description: Filtrar por nome (busca parcial)
          schema:
            type: string
        - name: email
          in: query
          description: Filtrar por email (busca parcial)
          schema:
            type: string
        - name: cpf
          in: query
          description: Filtrar por CPF (busca exata)
          schema:
            type: string
        - name: phone
          in: query
          description: Filtrar por telefone (busca parcial)
          schema:
            type: string
        - name: is_active
          in: query
          description: Filtrar por status ativo/inativo
          schema:
            type: boolean
        - name: from_birth_date
          in: query
          description: Data de nascimento inicial (formato YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to_birth_date
          in: query
          description: Data de nascimento final (formato YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: from_created_at
          in: query
          description: Data de criação inicial (formato YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to_created_at
          in: query
          description: Data de criação final (formato YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: sort_by
          in: query
          description: Campo para ordenação
          schema:
            type: string
            enum: [name, email, cpf, birth_date, created_at]
            default: created_at
        - name: sort_order
          in: query
          description: Direção da ordenação
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: per_page
          in: query
          description: Quantidade de resultados por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 15
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Client'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthenticated'

    post:
      tags:
        - Clients
      summary: Create client
      description: Cria um novo cliente
      security:
        - sanctum: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInput'
      responses:
        '201':
          description: Cliente criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          description: Dados de validação inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /api/clients/{id}:
    get:
      tags:
        - Clients
      summary: Get client
      description: Retorna os dados de um cliente específico
      security:
        - sanctum: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do cliente
          schema:
            type: integer
      responses:
        '200':
          description: Dados do cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          description: Cliente não encontrado

    put:
      tags:
        - Clients
      summary: Update client
      description: Atualiza os dados de um cliente
      security:
        - sanctum: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do cliente
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInput'
      responses:
        '200':
          description: Cliente atualizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          description: Cliente não encontrado
        '422':
          description: Dados de validação inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      tags:
        - Clients
      summary: Delete client
      description: Remove um cliente
      security:
        - sanctum: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do cliente
          schema:
            type: integer
      responses:
        '204':
          description: Cliente removido com sucesso
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          description: Cliente não encontrado

components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação Laravel Sanctum

  schemas:
    Client:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: João Silva
        email:
          type: string
          format: email
          example: joao.silva@example.com
        cpf:
          type: string
          pattern: '^\d{11}$'
          example: '12345678901'
        birth_date:
          type: string
          format: date
          example: '1990-05-15'
        phone:
          type: string
          example: '11987654321'
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00.000000Z'

    ClientInput:
      type: object
      required:
        - name
        - email
        - cpf
        - password
        - birth_date
        - phone
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
          example: João Silva
        email:
          type: string
          format: email
          maxLength: 255
          example: joao.silva@example.com
        cpf:
          type: string
          pattern: '^\d{11}$'
          description: CPF válido com 11 dígitos
          example: '12345678901'
        password:
          type: string
          minLength: 8
          maxLength: 255
          format: password
          example: password123
        birth_date:
          type: string
          format: date
          description: Data de nascimento (YYYY-MM-DD)
          example: '1990-05-15'
        phone:
          type: string
          maxLength: 20
          example: '11987654321'
        is_active:
          type: boolean
          default: true
          example: true

    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email field is required.
            cpf:
              - The cpf field is required.

    PaginationLinks:
      type: object
      properties:
        first:
          type: string
          example: http://localhost:8000/api/clients?page=1
        last:
          type: string
          example: http://localhost:8000/api/clients?page=10
        prev:
          type: string
          nullable: true
          example: null
        next:
          type: string
          nullable: true
          example: http://localhost:8000/api/clients?page=2

    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        from:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 10
        path:
          type: string
          example: http://localhost:8000/api/clients
        per_page:
          type: integer
          example: 15
        to:
          type: integer
          example: 15
        total:
          type: integer
          example: 150

  responses:
    Unauthenticated:
      description: Token de autenticação inválido ou ausente
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthenticated.